refactor(vcs): one PR composer for titlebar and explorer (D4)
Step 7 of the VCS unification (AD-024). The two near-copy composers
(titlebar GitPullRequestComposer, explorer PullRequestSection) are replaced
by a single src/components/git/PullRequestComposer with:

- one gating rule: no remote → publish hint; no non-default branch →
  feature-branch hint
- one source-branch seed preference (current branch in the Ship deck,
  active push branch in the explorer)
- one target-branch input (combo box with branch suggestions)
- shared preview/draft/create logic incl. the drafted-with-model notice

Hosts provide their own containers (Ship deck section, VcsSection).

Plan: docs/features/vcs-unification.md (Phase 3, step 7)
