+++ apps/desktop/src/components/apps/explorer/vcs/VcsPanel.tsx
@@ -13,7 +13,9 @@ import { useWorkspaceVcs, useVcsStore } from '@/stores/vcs';
 import { WorkingCopySection } from './WorkingCopySection';
 import { BranchesSection } from './BranchesSection';
 import { GitHubAuthBanner } from './GitHubAuthBanner';
-import { PullRequestSection } from './PullRequestSection';
+import { VcsSection } from './VcsSection';
+import { PullRequestComposer } from '@/components/git/PullRequestComposer';
+import { Github } from 'lucide-react';
 import { CommitLog } from './CommitLog';
 import { RemotesSection } from './RemotesSection';
 
@@ -110,11 +112,21 @@ export function VcsPanel({ workspaceId, onOpenDiff }: VcsPanelProps) {
 
             <GitHubAuthBanner className="mx-2" />
 
-            <PullRequestSection
-              workspaceId={workspaceId}
-              branches={ws?.branches ?? []}
-              activePushBranch={ws?.activePushBranch ?? null}
-            />
+            <VcsSection
+              title="Pull Request"
+              defaultOpen
+              badge={<Github className="ml-1 size-3 text-[var(--text-muted)]/70" />}
+            >
+              <div className="px-2 pb-2">
+                <PullRequestComposer
+                  compact
+                  workspaceId={workspaceId}
+                  hasRemote={(ws?.remotes.length ?? 0) > 0}
+                  preferredSourceBranch={ws?.activePushBranch ?? null}
+                  refreshKey={(ws?.branches ?? []).map((b) => b.name).sort().join('|')}
+                />
+              </div>
+            </VcsSection>
 
             <CommitLog
               workspaceId={workspaceId}
